Refactor : 당첨 숫자 예외 메소드를 WinningException.class로 이동
