+++ src/test/java/lotto/Exception/ExceptionControllerTest.java
@@ -17,13 +17,4 @@ public class ExceptionControllerTest {
 
 
 
-    @DisplayName("잘못된 당첨 숫자 예외 테스트")
-    @Test
-    void checkWinningNumberExceptionTest() {
-        Assertions.assertThatThrownBy(() ->  checkWinningNumberException(List.of(1,2,3,4,5,46))).isInstanceOf(IllegalArgumentException.class);
-        Assertions.assertThatThrownBy(() ->  checkWinningNumberException(List.of(1,2,3,4,5,0))).isInstanceOf(IllegalArgumentException.class);
-        Assertions.assertThatThrownBy(() ->  checkWinningNumberException(List.of(1,2,3,4,5))).isInstanceOf(IllegalArgumentException.class);
-        Assertions.assertThatThrownBy(() ->  checkWinningNumberException(List.of(1,2,3,4,5,5))).isInstanceOf(IllegalArgumentException.class);
-    }
-
 }
